+++ R/tbl_demographics.R
@@ -92,6 +92,18 @@ tbl_demographics <- function(data,
     structure(class = c("tbl_demographics", "tbl_summary", "gtsummary"))
 
   # return table ---------------------------------------------------------------
+  x$inputs$missing <- NULL
+  x$inputs$missing_stat <- NULL
+  x$inputs$missing_text <- NULL
+
+  x$inputs <- x$inputs |>
+    utils::modifyList(
+      val = list(
+        nonmissing = nonmissing,
+        nonmissing_stat = nonmissing_stat,
+        nonmissing_text = nonmissing_text
+      )
+    )
   x$call_list <- list(tbl_demographics = match.call())
   x
 }