+++ NEWS.md
@@ -1,3 +1,3 @@
 # crane 0.1.0.9000
 
-* Initial CRAN submission.
+* Update the `tbl_demographics()` returned `inputs` list to match the `tbl_demographics()`, whereas is returned the `tbl_summary()` inputs. (#31)